# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Shipment Tracking API
  version: 1.0.0
  description: API proxy for tracking shipments and retrieving tracking history.

servers:
  - url: /api/v1
    description: Current API version

paths:
  /tracking:
    get:
      summary: Track a shipment
      description: Retrieves tracking information for a specific courier and airway bill (tracking number).
      operationId: trackShipment
      tags:
        - Tracking
      parameters:
        - name: courier
          in: query
          required: true
          description: The courier code (e.g., jne, jnt, sicepat)
          schema:
            type: string
            pattern: "^[a-zA-Z]+$"
          example: jnt
        - name: tracking_number
          in: query
          required: true
          description: The airway bill (AWB) or tracking number
          schema:
            type: string
            pattern: "^[a-zA-Z0-9]+$"
          example: JP123456789
      responses:
        "200":
          description: Successful tracking response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackingResponseWrapper"
              example:
                status: "success"
                data:
                  summary:
                    awb: "JP123456789"
                    courier: "J&T Express"
                    service: "EZ"
                    status: "DELIVERED"
                    date: "2023-10-01 10:00:00"
                    desc: "Delivered to receiver"
                    amount: "15000"
                    weight: "1.5 kg"
                  detail:
                    origin: "JAKARTA"
                    destination: "BANDUNG"
                    shipper: "Toko Sukses"
                    receiver: "Budi"
                  history:
                    - date: "2023-10-01 10:00:00"
                      desc: "Delivered to receiver"
                      location: "BANDUNG"
                    - date: "2023-09-30 08:00:00"
                      desc: "Arrived at Drop Point"
                      location: "BANDUNG"
        "400":
          description: Bad request (validation error)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: "error"
                message: "invalid request payload: Field validation for 'Courier' failed on the 'required' tag"
        "404":
          description: Tracking number not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: "error"
                message: "Tracking number not found"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: "error"
                message: "Rate limit exceeded, please try again later"

  /history:
    get:
      summary: Get tracking history
      description: Retrieves paginated recent search history, optionally filtered by courier.
      operationId: getHistory
      tags:
        - History
      parameters:
        - name: page
          in: query
          required: false
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 1
        - name: limit
          in: query
          required: false
          description: Number of records per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          example: 10
        - name: courier
          in: query
          required: false
          description: Filter history by courier code
          schema:
            type: string
            pattern: "^[a-zA-Z]+$"
          example: jnt
      responses:
        "200":
          description: Successful history response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedHistoryResponse"
              example:
                status: "success"
                data:
                  data:
                    - id: 1
                      tracking_number: "JP123456789"
                      courier: "jnt"
                      searched_at: "2023-10-01T12:00:00Z"
                  total: 1
                  page: 1
                  limit: 10
                  total_pages: 1
        "400":
          description: Bad request (validation error)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: "error"
                message: "invalid pagination parameters"

components:
  schemas:
    TrackingResponseWrapper:
      type: object
      required:
        - status
        - data
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
        data:
          $ref: "#/components/schemas/TrackingData"

    PaginatedHistoryResponse:
      type: object
      required:
        - status
        - data
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
        data:
          $ref: "#/components/schemas/PaginatedData"

    ErrorResponse:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: string
          example: "error"
        message:
          type: string
          example: "An error occurred"

    TrackingData:
      type: object
      required:
        - summary
        - detail
        - history
      properties:
        summary:
          $ref: "#/components/schemas/TrackingSummary"
        detail:
          $ref: "#/components/schemas/TrackingDetail"
        history:
          type: array
          items:
            $ref: "#/components/schemas/TrackingHistoryItem"

    TrackingSummary:
      type: object
      properties:
        awb:
          type: string
        courier:
          type: string
        service:
          type: string
        status:
          type: string
        date:
          type: string
        desc:
          type: string
        amount:
          type: string
        weight:
          type: string

    TrackingDetail:
      type: object
      properties:
        origin:
          type: string
        destination:
          type: string
        shipper:
          type: string
        receiver:
          type: string

    TrackingHistoryItem:
      type: object
      properties:
        date:
          type: string
        desc:
          type: string
        location:
          type: string

    PaginatedData:
      type: object
      required:
        - data
        - total
        - page
        - limit
        - total_pages
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/HistoryRecord"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        total_pages:
          type: integer

    HistoryRecord:
      type: object
      required:
        - id
        - tracking_number
        - courier
        - searched_at
      properties:
        id:
          type: integer
        tracking_number:
          type: string
        courier:
          type: string
        searched_at:
          type: string
          format: date-time